chore: move pages without versions
