+++ web/apps/benchmarks-and-reports/src/app/(benchmarks-and-reports)/[version]/page.tsx
@@ -6,8 +6,12 @@ import { REPORTS } from "./_utils/constants";
 export default function ReportsPage({ params }) {
   return (
     <div className="container grid max-w-screen-3xl grid-cols-1 gap-4 pt-4 lg:grid-cols-2">
-      {REPORTS.map(({ label, href, description }, index) => (
-        <Link key={href} href={`/${params.version}/${href}`} className="group transition-opacity hover:opacity-70">
+      {REPORTS.map(({ label, href, description, showVersionSelect }, index) => (
+        <Link
+          key={href}
+          href={`${showVersionSelect ? `/${params.version}` : ""}${href}`}
+          className="group transition-opacity hover:opacity-70"
+        >
           <Card className="group-hover:-translate-y-1 flex h-full min-h-44 w-full flex-col items-center justify-between gap-1 px-8 py-4 shadow-sm transition-transform md:flex-row md:gap-12 dark:bg-neutral-900">
             <div>
               <CardTitle className="text-xl">{label}</CardTitle>
